Move field_info to taichi_core target
