+++ taichi/ui/backends/vulkan/renderables/mesh.h
@@ -18,7 +18,7 @@
 #include "taichi/ui/backends/vulkan/app_context.h"
 #include "taichi/ui/backends/vulkan/swap_chain.h"
 #include "taichi/ui/backends/vulkan/renderable.h"
-#include "taichi/ui/common/field_info.h"
+#include "taichi/program/field_info.h"
 #include "taichi/ui/backends/vulkan/scene.h"
 
 TI_UI_NAMESPACE_BEGIN
