+++ taichi/ui/common/renderable_info.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "taichi/ui/backends/vulkan/vertex.h"
-#include "taichi/ui/common/field_info.h"
+#include "taichi/program/field_info.h"
 #include "taichi/ui/utils/utils.h"
 
 TI_UI_NAMESPACE_BEGIN
